嘛共The拆開
